requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
astral==2.2
    # via -r requirements.in
cachetools==4.2.2
    # via google-auth
certifi==2021.5.30
    # via
    #   requests
    #   urllib3
cffi==1.14.6
    # via
    #   cryptography
    #   sounddevice
charset-normalizer==2.0.3
    # via requests
click==8.0.1
    # via -r requirements.in
cryptography==3.4.7
    # via
    #   pyopenssl
    #   urllib3
google-assistant-grpc==0.3.0
    # via -r requirements.in
google-auth==1.34.0
    # via google-auth-oauthlib
google-auth-oauthlib==0.4.5
    # via -r requirements.in
googleapis-common-protos==1.53.0
    # via google-assistant-grpc
grpcio==1.39.0
    # via google-assistant-grpc
idna==3.2
    # via
    #   requests
    #   urllib3
oauthlib==3.1.1
    # via requests-oauthlib
protobuf==3.17.3
    # via googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.20
    # via cffi
pyopenssl==20.0.1
    # via urllib3
pytz==2021.1
    # via astral
requests==2.26.0
    # via requests-oauthlib
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   grpcio
    #   protobuf
    #   pyopenssl
sounddevice==0.4.2
    # via -r requirements.in
speechrecognition==3.8.1
    # via -r requirements.in
tenacity==8.0.1
    # via -r requirements.in
timesched==1.6
    # via -r requirements.in
urllib3[secure]==1.26.6
    # via
    #   -r requirements.in
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
